improve error handling
now properly stops the binary if disconnection on either the websocket
or mqtt side of things occurs. The expectation is that the system would
restart the binary when the connectivity recovers (think: systemd with
restarting units.)

This will re-discover all the mower information, but that is required
anyway because we cannot assume no events are lost in-between had
websocket connection failure ocurred.